Moved file chooser to top of ui and hid content
Everything that's not the textfield and button for file chooser is hidden until a valid file is specified. The file chooser was moved to the top of the ui for better user experience, as the user could have filled out all of the colors then selected a file and those colors would be gone.